productive
* All methods now work only through soap;
* To delete references to usaepay class library
* The updated documentation (links to usaepay docs in header methods make Payment(CC data) and other)

+++ UsaepayHelper/IScedulePaymentInfo.cs
@@ -12,6 +12,6 @@ namespace KlikNPayUsaEPay
         string ExpirationDate { get; set; }
         string BillingAddress { get; set; }
 
-        decimal Amount { get; set; }
+        double Amount { get; set; }
     }
 }
